Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
pydantic-settings==2.0.3
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
nltk==3.8.1
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.0.1
alembic==1.12.1
psycopg2-binary=2.9.10
